option and escape sequences to disable bidi (#592, #392, ~#605)

+++ src/win.h
@@ -22,7 +22,7 @@ extern void win_reconfig(void);
 extern void win_update(void);
 extern void win_schedule_update(void);
 
-extern void win_text(int x, int y, wchar *text, int len, cattr attr, cattr *textattr, int lattr, bool has_rtl);
+extern void win_text(int x, int y, wchar *text, int len, cattr attr, cattr *textattr, ushort lattr, bool has_rtl);
 extern void win_update_mouse(void);
 extern void win_capture_mouse(void);
 extern void win_bell(config *);
